Machine-Learning-Models/ANN_azure_API/requirements.txt
jsonschema==3.0.1
lxml==4.3.0
numpy==1.16.2
pandas==0.24.1
xlrd==1.2.0
XlsxWriter==1.1.5
psutil==5.6.1
sqlalchemy==1.3.1
mysql-connector-python==8.0.15
pymongo==3.7.2
pyxlsb==1.0.5
azure-storage-blob==1.5.0
scikit-learn==0.22.2
tensorflow
matplotlib==3.0.2
seaborn
mlxtend